src/requirements.txt
annotated-types~=0.7.0
anyio~=4.7.0
certifi~=2024.8.30
distro~=1.9.0
exceptiongroup~=1.2.2
h11~=0.14.0
httpcore~=1.0.7
httpx~=0.28.1
idna~=3.10
isodate~=0.7.2
jiter~=0.8.0
openai~=1.57.1
psycopg2-binary~=2.9.7
pydantic~=2.10.3
pydantic_core~=2.27.1
python-dotenv~=1.0.1
python-telegram-bot~=21.7
sniffio~=1.3.1
tqdm~=4.67.1
typing_extensions~=4.12.2